@ahsonkhan 2019-04-05
> The "after" codegen is also 6 bytes smaller than the "before" codegen.

Nice! :)